chore: apply formatting fixes and config setup
Applied pre-commit hooks to enforce consistent code style and
formatting across the project. Added configuration files and
updated existing files for whitespace and formatting compliance.

Includes .pre-commit-config.yaml and .gitmessage setup.
