+++ .pre-commit-config.yaml
@@ -1,0 +1,32 @@
+repos:
+-   repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v5.0.0  # Use the ref you want to point at
+    hooks:
+    -   id: check-toml
+    -   id: end-of-file-fixer
+    -   id: check-added-large-files
+-   repo: https://github.com/codespell-project/codespell
+    rev: v2.4.1
+    hooks:
+    - id: codespell
+-   repo: https://github.com/doublify/pre-commit-rust
+    rev: v1.0
+    hooks:
+    -   id: fmt
+    -   id: cargo-check
+-   repo: local
+    hooks:
+    # -   id: cargo-test
+    #     name: cargo test
+    #     description: Run tests
+    #     entry: cargo test
+    #     language: system
+    #     types: [rust]
+    #     pass_filenames: false
+    -   id: cargo-clippy
+        name: cargo clippy
+        description: Run clippy linter
+        entry: cargo clippy -- --deny unused_variables
+        language: system
+        types: [rust]
+        pass_filenames: false
