+++ astro/Cargo.toml
@@ -7,7 +7,7 @@ authors.workspace = true
 repository.workspace = true
 
 [lib]
-crate-type = ["dylib"]
+crate-type = ["dylib","rlib"]
 
 [dependencies]
 bumpalo = {version="3.17.0",features=["boxed"]}
