+++ physim-core/src/bin/messages.rs
@@ -7,7 +7,7 @@ use std::{
 };
 
 use physim_core::{
-    messages::{callback, Message, MessageBus, MessageClient, MessagePriority},
+    messages::{Message, MessageBus, MessageClient, MessagePriority, callback},
     plugin::{set_bus, transform::TransformElementHandler},
 };
